dicom-matlab-poc/requirements.txt
pydicom>=2.4.0
numpy>=1.24.0
Pillow>=10.0.0
